♻️ UI와 로직 분리
UI 부분은 template에, 로직은 page로 분리
